# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: INNDOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [rent, sale, bnb, hotel, hostel]
        - name: search
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: string
        - name: minLat
          in: query
          schema:
            type: number
        - name: maxLat
          in: query
          schema:
            type: number
        - name: minLng
          in: query
          schema:
            type: number
        - name: maxLng
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyInput"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /properties/{id}/availability:
    get:
      operationId: getPropertyAvailability
      tags: [properties]
      summary: Get booked date ranges for a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of booked date ranges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookedRange"
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [rent, sale, bnb, hotel, hostel]
        price:
          type: integer
        address:
          type: string
        beds:
          type: integer
        baths:
          type: integer
        sqft:
          type: integer
        guests:
          type: integer
          nullable: true
        image:
          type: string
        images:
          type: array
          items:
            type: string
        videos:
          type: array
          items:
            type: string
        details:
          type: object
          additionalProperties: true
        isVerified:
          type: boolean
        tags:
          type: array
          items:
            type: string
        lat:
          type: string
          nullable: true
        lng:
          type: string
          nullable: true
        createdAt:
          type: string
        description:
          type: string
          nullable: true
        subtype:
          type: string
          nullable: true
        hourlyRate:
          type: integer
          nullable: true
        priceUnit:
          type: string
          nullable: true
        totalUnits:
          type: integer
          minimum: 1
        propertyStatus:
          type: string
          enum: [pending, approved, flagged, sold]
        activeBookingsCount:
          type: integer
        ownerPhone:
          type: string
          nullable: true
        ownerEmail:
          type: string
          nullable: true
        ownerAvatar:
          type: string
          nullable: true
        ownerBusinessName:
          type: string
          nullable: true
        savedAt:
          type: string
          nullable: true
        ownerName:
          type: string
          nullable: true
      required:
        - id
        - ownerId
        - title
        - type
        - price
        - address
        - beds
        - baths
        - sqft
        - image
        - images
        - videos
        - details
        - isVerified
        - tags
        - totalUnits
    CreatePropertyInput:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [rent, sale, bnb, hotel, hostel]
        price:
          type: integer
        address:
          type: string
        beds:
          type: integer
        baths:
          type: integer
        sqft:
          type: integer
        guests:
          type: integer
        image:
          type: string
        images:
          type: array
          items:
            type: string
        videos:
          type: array
          items:
            type: string
        details:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        lat:
          type: string
        lng:
          type: string
        subtype:
          type: string
        hourlyRate:
          type: integer
        priceUnit:
          type: string
        totalUnits:
          type: integer
          minimum: 1
      required:
        - title
        - type
        - price
        - address
...
    BookedRange:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [pending, confirmed, blocked]
      required:
        - startDate
        - endDate
        - status
...