# lib/api-zod/src/generated/types/createPropertyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * INNDOS API specification
 * OpenAPI spec version: 0.2.0
 */
import type { CreatePropertyInputDetails } from "./createPropertyInputDetails";
import type { CreatePropertyInputType } from "./createPropertyInputType";

export interface CreatePropertyInput {
  title: string;
  type: CreatePropertyInputType;
  price: number;
  address: string;
  beds?: number;
  baths?: number;
  sqft?: number;
  guests?: number;
  image?: string;
  images?: string[];
  videos?: string[];
  details?: CreatePropertyInputDetails;
  tags?: string[];
  description?: string | null;
  lat?: string;
  lng?: string;
  subtype?: string;
  hourlyRate?: number;
  priceUnit?: string;
  /** @minimum 1 */
  totalUnits?: number;
}